deconnexion

+++ Views/liste-articles.php
@@ -11,8 +11,12 @@
             <h5 class="card-title"><?= $article->titre ?></h5>
             <p class="card-text"><?= $article->contenu ?></p>
             <a href="index.php?route=details&id=<?= $article->id ?>" class="btn btn-primary"><i class="bi bi-book"></i> Lire</a> <br>
-            <a href="index.php?route=modifier&id=<?= $article->id ?>"><i class="bi bi-eraser"></i> Modifier</a>
-            <a href="index.php?route=supprimer&id=<?= $article->id ?>" class="text-danger"><i class="bi bi-trash"></i> Supprimer</a>
+            
+        
+            <?php if (empty($_SESSION['identifiant']) && $_SESSION['identifiant'] == 'admin') : ?>
+                <a href="index.php?route=modifier&id=<?= $article->id ?>"><i class="bi bi-eraser"></i> Modifier</a>
+                <a href="index.php?route=supprimer&id=<?= $article->id ?>" class="text-danger" onclick="return confirm('Etes-vous sûr de vouloir supprimer l\'article ?')"><i class="bi bi-trash"></i> Supprimer</a>
+            <?php endif; ?>
         </div>
     </div>
     <?php } ?>
